+++ gpu/a_star_gpu_3.c
@@ -23,7 +23,7 @@ __host__ __device__ int manhattanDistance(int x1, int y1, int x2, int y2) {
 
 // GPU
 __global__ void MultiFrontierExpansion(int* bucket_nodes_d, int* bucket_sizes_d, int currentBucket, int currentBucketSize,
-int* gScore_d, int* parent_d, int gridSize_x, int gridSize_y, float* gridData_d, int goalIndex_x, int goalIndex_y) {
+int* gScore_d, int* parent_d, int gridSize_x, int gridSize_y, int* gridData_d, int goalIndex_x, int goalIndex_y) {
     int thread_id = blockIdx.x * blockDim.x + threadIdx.x;
     if (thread_id >= currentBucketSize) return;
 
@@ -37,7 +37,7 @@ int* gScore_d, int* parent_d, int gridSize_x, int gridSize_y, float* gridData_d,
         int neighborIndex = neighborIndex_y * gridSize_x + neighborIndex_x;
 
         if (neighborIndex_x < 0 || neighborIndex_x >= gridSize_x || neighborIndex_y < 0 || neighborIndex_y >= gridSize_y) continue;
-        if (gridData_d[neighborIndex] == 2.0f) continue;
+        if (gridData_d[neighborIndex] == 2) continue;
 
         int newVal = gScore_d[index] + 1;
         int oldVal = atomicMin(&gScore_d[neighborIndex], newVal);
@@ -78,7 +78,7 @@ void Run_AStar(Grid_2D_Device* grid, int startIndex_x, int startIndex_y, int goa
     for (int i = 0; i < gridSize; i++) gScore[i] = INT_MAX; // set each byte to maximum value
     gScore[startIndex] = 0;
 
-    // No need to set grid elements to 1.0f anymore
+    // No need to set grid elements to 1 anymore
     // No f-score for now (search dijkstra-style)
 
     int* gScore_d;
@@ -89,9 +89,9 @@ void Run_AStar(Grid_2D_Device* grid, int startIndex_x, int startIndex_y, int goa
     cudaMalloc((void**)&parent_d, sizeof(int)*gridSize);
     cudaMemcpy(parent_d, grid->parent, sizeof(int)*gridSize, cudaMemcpyHostToDevice);
 
-    float* gridData_d;
-    cudaMalloc((void**)&gridData_d, sizeof(float)*gridSize);
-    cudaMemcpy(gridData_d, grid->data, sizeof(float)*gridSize, cudaMemcpyHostToDevice);
+    int* gridData_d;
+    cudaMalloc((void**)&gridData_d, sizeof(int)*gridSize);
+    cudaMemcpy(gridData_d, grid->data, sizeof(int)*gridSize, cudaMemcpyHostToDevice);
 
     int* bucket_sizes_d;
     cudaMalloc((void**)&bucket_sizes_d, sizeof(int)*NUM_BUCKETS);
@@ -180,11 +180,11 @@ int main(int argc, char* argv[]) {
     cudaMemcpyToSymbol(offset_x, offset_xh, sizeof(int) * 4, 0, cudaMemcpyHostToDevice);
     cudaMemcpyToSymbol(offset_y, offset_yh, sizeof(int) * 4, 0, cudaMemcpyHostToDevice);
 
-    Grid_2D_Device* myGrid = CreateGrid(4, 4, 0.0f);
+    Grid_2D_Device* myGrid = CreateGrid(4, 4, 0);
     
-    myGrid->data[4] = 2.0f;
-    myGrid->data[5] = 2.0f;
-    myGrid->data[6] = 2.0f;
+    myGrid->data[4] = 2;
+    myGrid->data[5] = 2;
+    myGrid->data[6] = 2;
 
     Run_AStar(myGrid, 0, 0, 0, 2);
